Start moving the "runtime" parts of the client code into an *.so
The client portion of the DD WebGPU runtime will be shipped as a shared
library pre-installed in the DD VM image. The main goal of this library
is to provide all relevant WebGPU symbols and a minimal set of calls
needed to initialise the runtime. So start separating the client init
code into a shared library and make sure that library contains all of
libdawn_procs.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>